requirements.txt
python-scribd
